requirements-dev.txt
model-mommy==2.0.0
requests_mock==1.8.0
django-debug-toolbar==2.2